+++ tests/async_engine/test_chat_template.py
@@ -3,7 +3,7 @@ import pathlib
 
 import pytest
 
-from vllm.entrypoints.openai.chat_utils import load_chat_template
+from vllm.entrypoints.chat_utils import load_chat_template
 from vllm.entrypoints.openai.protocol import ChatCompletionRequest
 from vllm.transformers_utils.tokenizer import get_tokenizer
