+++ vllm/entrypoints/openai/serving_tokenization.py
@@ -2,9 +2,9 @@ from typing import List, Optional
 
 from vllm.config import ModelConfig
 from vllm.engine.async_llm_engine import AsyncLLMEngine
-from vllm.entrypoints.openai.chat_utils import (ConversationMessage,
-                                                load_chat_template,
-                                                parse_chat_message_content)
+from vllm.entrypoints.chat_utils import (ConversationMessage,
+                                         load_chat_template,
+                                         parse_chat_message_content)
 from vllm.entrypoints.openai.protocol import (DetokenizeRequest,
                                               DetokenizeResponse,
                                               TokenizeRequest,
